+++ frontend/frontend_2/src/components/views/DocumentView/DocumentModal.js
